+++ source1/data_structures/source_shared.py
@@ -2,7 +2,7 @@ import math
 import random
 import struct
 
-from ..byte_io_mdl import ByteIO
+from ...byte_io_mdl import ByteIO
 
 
 class SourceVector:
