+++ source1/data_structures/vvd_data.py
@@ -1,7 +1,7 @@
 import traceback
 from typing import List
 
-from ..byte_io_mdl import ByteIO
+from ...byte_io_mdl import ByteIO
 from ..data_structures.source_shared import SourceVertex
