+++ source1/mdl/mdl_readers/mdl_v10.py
@@ -1,11 +1,11 @@
 from typing import List
 
-from ...byte_io_mdl import ByteIO
+from ....byte_io_mdl import ByteIO
 from ...data_structures.mdl_data import SourceMdlFileData, SourceMdlBone, SourceMdlAttachment, SourceMdlBodyPart, \
     SourceMdlTexture, SourceMdlModel, StudioHDRFlags, SourceMdlMesh, SourceBase
 from ...data_structures.source_shared import SourceVector
 from ...mdl.mdl_readers.mdl_v49 import SourceMdlFile49
-from ...utilities.progressbar import ProgressBar
+from ....utilities.progressbar import ProgressBar
 
 
 class SourceMdlFile10(SourceMdlFile49):
