+++ source1/mdl/mdl_readers/mdl_v53.py
@@ -1,4 +1,4 @@
-from ...byte_io_mdl import ByteIO
+from ....byte_io_mdl import ByteIO
 from ...data_structures.mdl_data import SourceMdlFileData
 from ...mdl.mdl_readers.mdl_v49 import SourceMdlFile49
