+++ source1/mdl/source_model.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from ..byte_io_mdl import ByteIO
+from ...byte_io_mdl import ByteIO
 from ..mdl import qc_generator
 from ..mdl.mdl_readers.mdl_v10 import SourceMdlFile10
 from ..mdl.mdl_readers.mdl_v48 import SourceMdlFile48
@@ -8,7 +8,7 @@ from ..mdl.mdl_readers.mdl_v49 import SourceMdlFile49
 from ..mdl.mdl_readers.mdl_v53 import SourceMdlFile53
 from ..mdl.vtx_readers.vtx_v7 import SourceVtxFile7
 from ..mdl.vvd_readers.vvd_v4 import SourceVvdFile4
-from ..utilities.path_utilities import case_insensitive_file_resolution
+from ...utilities.path_utilities import case_insensitive_file_resolution
 
 
 class SourceModel:
