+++ source1/vtf/vmt.py
@@ -1,6 +1,6 @@
 import os
 
-from ..utilities.valve_utils import GameInfoFile, KeyValueFile, MaterialPathResolver
+from ...utilities.valve_utils import GameInfoFile, KeyValueFile, MaterialPathResolver
 
 from pathlib import Path
